+++ indico/modules/events/registration/client/js/form/fields/registry.js
@@ -43,6 +43,11 @@ import SingleChoiceInput, {
 } from './SingleChoiceInput';
 import TextAreaInput, {TextAreaSettings} from './TextAreaInput';
 import TextInput, {TextSettings, textSettingsFormValidator} from './TextInput';
+import TimetableSessionsInput, {
+  TimetableSessionsSettings,
+  timetableSessionsSettingsInitialData,
+  sessionsSettingsFormValidator,
+} from './TimetableSessions';
 
 /*
 Available keys:
@@ -194,6 +199,17 @@ const fieldRegistry = {
     inputComponent: PictureInput,
     icon: 'image',
     settingsComponent: PictureSettings,
+  },
+  sessions: {
+    title: Translate.string('Timetable Sessions'),
+    inputComponent: TimetableSessionsInput,
+    settingsComponent: TimetableSessionsSettings,
+    settingsFormInitialData: timetableSessionsSettingsInitialData,
+    settingsFormValidator: sessionsSettingsFormValidator,
+    noRequired: true,
+    hasPrice: false,
+    icon: 'calendar-day',
+    renderAsFieldset: true,
   },
 };